% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSMVRead.R
\name{POSMVRead}
\alias{POSMVRead}
\title{Extract information from binary Applanix POSMV files and write to plain text}
\usage{
POSMVRead(
  input,
  output,
  recursive = TRUE,
  append = FALSE,
  pattern = "\\\\.\\\\d+",
  verbose = FALSE,
  tmpdir = tempdir()
)
}
\arguments{
\item{input}{Character vector of file names or a directory containing Applanix POSMV binary files}

\item{output}{Output file name for resultant csv file.}

\item{recursive}{logical (default is TRUE). Should the listing recurse into directories?}

\item{append}{Logical indicating whether the output should be appended to an existing output}

\item{pattern}{Pattern of file names to search for in directory. Default is to is a file extension of just numbers (e.g. '.000').}

\item{verbose}{logical indicating whether to print progress to display (deafult is FALSE)}

\item{tmpdir}{character vector giving the directory name for temporary files to be stored}
}
\description{
Extract information from binary Applanix POSMV files and write to plain text. Note this function requires python 3 to be installed on your computer.
}
\details{
This R function uses system calls to a python script written by Paul Kennedy (https://github.com/pktrigg/posmv). Data is then read into R to do some tidying so it can be exported as a clean csv.
}
